+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - none
 
 # v2025.5.1-beta
+- av - Rename CEDRIVER73, dbk32, dbk64, ManufacturerName.
 - ci - Upgrade ci to nodejs v22.
 
 # v2023.11.1-beta